modifiche nella visualizzazione e nella stampa dell'ordine

+++ AeronPrint/AeronPrint/OrderListCtrl.cpp
@@ -44,7 +44,7 @@ std::vector<int> OrderListCtrl::GetSelectedItems()
 	int current = -1;
 	std::vector<int> items;
 	
-	while (items.size() < GetSelectedItemCount())
+	while (static_cast<int>(items.size()) < GetSelectedItemCount())
 	{
 		current = GetNextItem(current, wxLIST_NEXT_ALL, wxLIST_STATE_SELECTED);
 		items.push_back(current);
